% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bobibob2-package.R
\docType{package}
\name{bobibob2-package}
\alias{bobibob2}
\alias{bobibob2-package}
\title{bobibob2: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nhenry50/bobibob2}
  \item Report bugs at \url{https://github.com/nhenry50/bobibob2/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Henry \email{nicolashenry50@gmail.com} (\href{https://orcid.org/{{orcid}}}{ORCID})

}
\keyword{internal}
