% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-gbif.R
\name{check_status}
\alias{check_status}
\title{Check status of download preparation}
\usage{
check_status()
}
\value{
Messages indicating the status of the download prep (preparing, running, succedded), then metadata about the download request once the preparation has succeeded.
}
\description{
Check status of download preparation
}
\examples{
\dontrun{
check_status()
}
}
