% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo-network.R
\name{data_for_focus}
\alias{data_for_focus}
\title{Generate a minimal table for the per taxo dashboards}
\usage{
data_for_focus(tax, dat)
}
\value{
A minimal data.table with MEOW info
}
\description{
Generate a minimal table for the per taxo dashboards
}
