% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo-network.R
\name{data_for_net}
\alias{data_for_net}
\title{Preparation of the raw data for the project}
\usage{
data_for_net()
}
\value{
A data.table extra columns for taxonomy
}
\description{
Preparation of the raw data for the project
}
