% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo-network.R
\name{data_for_net_nodes_edges}
\alias{data_for_net_nodes_edges}
\title{Node and edge data for the taxonomic network}
\usage{
data_for_net_nodes_edges(dat = data_for_net())
}
\value{
List of data.tables
}
\description{
Node and edge data for the taxonomic network
}
