% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-gbif.R
\name{gbif_download}
\alias{gbif_download}
\title{Download occurrence data from gbif}
\usage{
gbif_download(dl_key)
}
\arguments{
\item{dl_key}{The single identifier of the download prepared with the function gbif_download_init (The Download key printed by the function rgbif::occ_download_meta(res))}
}
\value{
a zip file containing the occurrence data imported from Gbif
}
\description{
Download occurrence data from gbif
}
\examples{
\dontrun{
gbif_download("0071104-210914110416597")
}

}
