% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-gbif.R
\name{gbif_download_init}
\alias{gbif_download_init}
\title{Prepare occurrence data download from Gbif}
\usage{
gbif_download_init()
}
\value{
Metadata about the download, including download status to check when the download is ready.
}
\description{
Prepare occurrence data download from Gbif
}
\examples{
\dontrun{
gbif_download_init()
}

}
