% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo-network.R
\name{taxo_network}
\alias{taxo_network}
\title{Generate a interactive taxonomic network}
\usage{
taxo_network(edge_and_node = data_for_net_nodes_edges())
}
\value{
A interactive taxonomic network
}
\description{
Generate a interactive taxonomic network
}
