% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bobibob2-package.R
\docType{package}
\name{bobibob2-package}
\alias{bobibob2}
\alias{bobibob2-package}
\title{bobibob2: A Dashboard to Visualise Macroalgae Biogeography}
\description{
This package will do amazing things... to be updated. URL: https://github.com/nhenry50/bobibob2
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/nhenry50/bobibob2/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Henry \email{nicolashenry50@gmail.com} (\href{https://orcid.org/{{0000-0002-7702-1382}}}{ORCID})

}
\keyword{internal}
