% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bobibob2-package.R
\docType{package}
\name{bobibob2-package}
\alias{bobibob2}
\alias{bobibob2-package}
\title{bobibob2: A Dashboard to Visualise Macroalgae Biogeography}
\description{
Allows exploration of the diversity and global distribution of all Kelp species, Laminaricaeae (those available in GBIF, Global Biodiversity Information Faciltity; https://www.gbif.org). Data is extracted directly from GBIF, determined by the desired taxa parameters - i.e. Laminaricaeae. The taxonomic structure is visualised as a network. Each node in the taxonomy network links to an html displaying a range of interactive figures, summarising occurance of the species of interest in time and space. For details on how to adapt this dashboard, including downloading desired taxa data see github. URL: https://github.com/nhenry50/bobibob2
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/nhenry50/bobibob2/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Henry \email{nicolashenry50@gmail.com} (\href{https://orcid.org/{{0000-0002-7702-1382}}}{ORCID})

}
\keyword{internal}
